+++ src/typeDefinition.ml
@@ -118,7 +118,7 @@ module Constructors = struct
       (case : Types.constructor_declaration)
       : (t * (RecordSkeleton.t * Name.t list * Type.t) option) Monad.t =
       let { Types.cd_args; cd_id; cd_loc; cd_res; _ } = case in
-      set_loc (Loc.of_location cd_loc) (
+      set_loc cd_loc (
       let* constructor_name =
         PathName.map_constructor_name
           (Ident.name cd_id) (Name.to_string typ_name) in
@@ -129,7 +129,7 @@ module Constructors = struct
         Type.of_typs_exprs true typ_vars param_typs >>= fun (param_typs, _, _) ->
         return (param_typs, None)
       | Cstr_record labeled_typs ->
-        set_loc (Loc.of_location cd_loc) (
+        set_loc cd_loc (
         (
           labeled_typs |>
           List.map (fun { Types.ld_type; _ } -> ld_type) |>
@@ -263,7 +263,7 @@ module Inductive = struct
 
   let get_notation_module_name (inductive : t) : SmartPrint.t =
     !^ "ConstructorRecords_" ^-^
-    separate (!^ "_") (inductive.typs |> List.map (fun (name, _, _) ->
+    separate (!^ "_") (inductive.constructor_records |> List.map (fun (name, _) ->
       Name.to_coq name
     ))
 
@@ -273,6 +273,7 @@ module Inductive = struct
     | constructor_records ->
       let notation_module_name = get_notation_module_name inductive in
       Some (
+        !^ "(** Records for the constructor parameters *)" ^^ newline ^^
         !^ "Module" ^^ notation_module_name ^-^ !^ "." ^^ newline ^^
         indent (
           separate newline (
@@ -344,12 +345,9 @@ module Inductive = struct
           } ->
           newline ^^ nest (
             !^ "|" ^^ Name.to_coq constructor_name ^^ !^ ":" ^^
-            (match typ_vars with
-            | [] -> empty
-            | _ ->
-              !^ "forall" ^^ braces (
-                separate space (typ_vars |> List.map Name.to_coq) ^^ !^ ":" ^^ Pp.set
-              ) ^-^ !^ ","
+            Name.to_coq_list_or_empty typ_vars (fun typ_vars ->
+              !^ "forall" ^^ braces (nest (typ_vars ^^ !^ ":" ^^ Pp.set)) ^-^
+              !^ ","
             ) ^^
             group @@ separate space (param_typs |> List.map (fun param_typ ->
               group (Type.to_coq (Some subst) (Some Type.Context.Arrow) param_typ ^^ !^ "->")
@@ -625,7 +623,7 @@ let of_ocaml (typs : type_declaration list) : t Monad.t =
       "We do not handle extensible types"
   | _ ->
     (typs |> Monad.List.fold_left (fun (constructor_records, notations, records, typs) typ ->
-      set_loc (Loc.of_location typ.typ_loc) (
+      set_loc typ.typ_loc (
       let* name = Name.of_ident false typ.typ_id in
       AdtParameters.of_ocaml typ.typ_type.type_params >>= fun typ_args ->
       match typ with