+++ tests/top_level_definition.ml
@@ -1,0 +1,8 @@
+let () =
+  let _ = 1 + 1 in
+  ignore 2;
+  ()
+
+module M = struct
+  ignore (1 + 1)
+end